+++ CHANGES.md
@@ -3,4 +3,5 @@
 ## Unreleased
 
 ## v0.0.0 - 2024-xx-xx
-
+- Implement `linksmith inventory` and `linksmith output-formats`
+  subcommands, based on `sphobjinv` and others. Thanks, @bskinn.